Implement the wrapper around named pipes
Adds an implementation for a socket based off the general idea described in ymnk/jsch-agent-proxy#34

This general approach seems to work when running similar code using the
JVM, yet seems to fail when used in conjunction with Cyberduck:
```
java.io.FileNotFoundException: 'FileStream was asked to open a device that was not a file. For support for devices like 'com1:' or 'lpt1:', call CreateFile, then use the FileStream constructors that take an OS handle as an IntPtr.'

This exception was originally thrown at this call stack:
    java.io.FileDescriptor.open(string, int, int)
    java.io.FileDescriptor.openReadWrite(string)
    java.io.RandomAccessFile.open(string, int)
    java.io.RandomAccessFile.RandomAccessFile(java.io.File, string)
    java.io.RandomAccessFile.RandomAccessFile(string, string)
    ch.cyberduck.core.sftp.openssh.RandomAccessFileSocketFactory.WindowsSocket.WindowsSocket(ch.cyberduck.core.sftp.openssh.RandomAccessFileSocketFactory, string) in OpenSSHAgentAuthenticator.java
    ch.cyberduck.core.sftp.openssh.RandomAccessFileSocketFactory.open(string) in OpenSSHAgentAuthenticator.java
    com.jcraft.jsch.agentproxy.connector.SSHAgentConnector.open() in SSHAgentConnector.java
    com.jcraft.jsch.agentproxy.connector.SSHAgentConnector.SSHAgentConnector(com.jcraft.jsch.agentproxy.USocketFactory, string) in SSHAgentConnector.java
```
Some Googling indicates that this likely originates from C#,
so an alternate approach -- or a fix on the C# side -- may be necessary.